src/ingestion/vision.py
"""
Vision processing module for extracting and analyzing diagrams in documents.
"""
import io
from typing import Dict, List, Any, Tuple
import base64

from google.cloud import vision_v1 as vision
from google.cloud import documentai_v1 as documentai
import numpy as np
from PIL import Image

# Initialize Vision client
vision_client = vision.ImageAnnotatorClient()


def extract_diagrams_from_pdf(pdf_content: bytes) -> List[Dict[str, Any]]:
    """
    Extract diagrams from a PDF document.
    
    Args:
        pdf_content: The raw PDF content
    
    Returns:
        List of dictionaries containing diagrams with their metadata
    """
    # For PDF processing, we'd typically:
    # 1. Convert PDF pages to images
    # 2. Process each image with Vision AI
    # 3. Identify diagrams in each page
    
    # This is a simplified implementation that would need to be expanded
    # with PDF library like PyPDF2 or pdf2image in production
    
    # Placeholder for extracted diagrams
    diagrams = []
    
    # TODO: In production, convert PDF to images and process each page
    # For now, we'll simulate finding diagrams on pages 1 and 2
    
    # This would be replaced with actual image extraction from PDF
    mock_diagram1 = {
        "image_content": pdf_content[:1000],  # First 1000 bytes as mock image
        "page": 1,
        "bbox": {
            "x1": 0.1,
            "y1": 0.2,
            "x2": 0.8,
            "y2": 0.7
        }
    }
    
    mock_diagram2 = {
        "image_content": pdf_content[1000:2000],  # Next 1000 bytes as mock image
        "page": 2,
        "bbox": {
            "x1": 0.2,
            "y1": 0.3,
            "x2": 0.9,
            "y2": 0.8
        }
    }
    
    diagrams.append(mock_diagram1)
    diagrams.append(mock_diagram2)
    
    return diagrams


def identify_diagrams(
    objects: List[vision.LocalizedObjectAnnotation], 
    text_annotation: vision.TextAnnotation
) -> List[Dict[str, Any]]:
    """
    Identify diagrams in an image based on object detection and text analysis.
    
    Args:
        objects: Objects detected in the image
        text_annotation: Text detected in the image
    
    Returns:
        List of dictionaries containing diagram information
    """
    diagrams = []
    
    # In a real implementation, we would:
    # 1. Use object detection to find diagram-like elements
    # 2. Check text nearby for captions like "Figure X", "Diagram", etc.
    # 3. Use layout analysis to identify technical drawings
    
    # Identify objects that might be diagrams (simplified)
    diagram_objects = [obj for obj in objects if _is_likely_diagram(obj)]
    
    for i, obj in enumerate(diagram_objects):
        # Extract bounding box
        bbox = {
            "x1": obj.bounding_poly.normalized_vertices[0].x,
            "y1": obj.bounding_poly.normalized_vertices[0].y,
            "x2": obj.bounding_poly.normalized_vertices[2].x,
            "y2": obj.bounding_poly.normalized_vertices[2].y,
        }
        
        # In production, we'd extract the actual image content
        # For this example, we'll create a placeholder
        diagram = {
            "image_content": b"placeholder_binary_content",  # In production: extract this region from image
            "bbox": bbox,
            "confidence": obj.score,
            "object_type": obj.name
        }
        
        diagrams.append(diagram)
    
    return diagrams


def _is_likely_diagram(obj: vision.LocalizedObjectAnnotation) -> bool:
    """
    Determine if an object is likely to be a diagram or technical drawing.
    
    Args:
        obj: The detected object
    
    Returns:
        Boolean indicating if the object is likely a diagram
    """
    # Simplified logic - in production, this would be more sophisticated
    diagram_related_objects = [
        "diagram", "chart", "drawing", "graph", "figure", 
        "blueprint", "schematic", "technical drawing"
    ]
    
    # Check if object name is in our list
    if any(term in obj.name.lower() for term in diagram_related_objects):
        return True
    
    # Check confidence score (diagrams often have distinct boundaries)
    if obj.score > 0.7:
        # Additional checks could examine the shape, content, etc.
        return True
    
    return False


def extract_image_from_region(
    image_content: bytes, 
    bbox: Dict[str, float]
) -> bytes:
    """
    Extract a region from an image based on a bounding box.
    
    Args:
        image_content: The full image content
        bbox: Normalized coordinates (x1,y1,x2,y2) for the bounding box
    
    Returns:
        Binary content of the extracted region
    """
    # Convert image content to PIL Image
    image = Image.open(io.BytesIO(image_content))
    
    # Get actual pixel coordinates from normalized coordinates
    width, height = image.size
    x1 = int(bbox["x1"] * width)
    y1 = int(bbox["y1"] * height)
    x2 = int(bbox["x2"] * width)
    y2 = int(bbox["y2"] * height)
    
    # Crop image to the specified region
    region = image.crop((x1, y1, x2, y2))
    
    # Convert back to bytes
    output = io.BytesIO()
    region.save(output, format=image.format or 'PNG')
    
    return output.getvalue()
